+++ packages/styles/src/GlobalStyle/index.ts
@@ -1,1 +1,1 @@
-export { default as GlobalStyle } from './GlobalStyle';
+export { default } from './GlobalStyle';