+++ packages/styles/src/index.ts
@@ -1,4 +1,5 @@
-export * from './GlobalStyle';
+export { default as GlobalStyle } from './GlobalStyle';
+export { default as ThemeProvider } from './ThemeProvider';
 
 export * from './const';
 export * from './mixins';